starting to separate dataloader from the training code
