+++ python/HEPCNN/keras_model_default.py
