+++ python/HEPCNN/torch_model_default.py
